@@ -5,7 +5,7 @@ publish_to: none
 
 environment:
   sdk: ">=3.3.3 <4.0.0"
-  flutter: "3.19.6"
+  flutter: "3.24.1"
 
 dependencies:
   flutter:
