@@ -1,3 +1,4 @@
+import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/drawing_creator.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/drawing_parts.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/drawing_tools/data_model/edge_point.dart';
@@ -7,13 +8,14 @@ import 'package:flutter/material.dart';
 /// Creates a Line drawing piece by piece collected on every gesture
 /// exists in a widget tree starting from selecting a line drawing tool and
 /// until drawing is finished
-class LineDrawingCreator extends DrawingCreator<LineDrawing> {
+class LineDrawingCreator extends DrawingCreator<Drawing> {
   /// Initializes the line drawing creator.
   const LineDrawingCreator({
-    required OnAddDrawing<LineDrawing> onAddDrawing,
+    required OnAddDrawing<Drawing> onAddDrawing,
     required double Function(double) quoteFromCanvasY,
     required this.clearDrawingToolSelection,
     required this.removeUnfinishedDrawing,
+    required this.createLineDrawing,
     Key? key,
   }) : super(
           key: key,
@@ -27,11 +29,21 @@ class LineDrawingCreator extends DrawingCreator<LineDrawing> {
   /// Callback to remove unfinished drawing from the list of drawings.
   final VoidCallback removeUnfinishedDrawing;
 
+  /// Callback which creates a [LineDrawing] or [LineDrawingMobile]
+  /// with the specified parameters.
+  final Drawing Function({
+    required DrawingParts drawingPart,
+    EdgePoint startEdgePoint,
+    EdgePoint endEdgePoint,
+    bool exceedStart,
+    bool exceedEnd,
+  }) createLineDrawing;
+
   @override
-  DrawingCreatorState<LineDrawing> createState() => _LineDrawingCreatorState();
+  DrawingCreatorState<Drawing> createState() => _LineDrawingCreatorState();
 }
 
-class _LineDrawingCreatorState extends DrawingCreatorState<LineDrawing> {
+class _LineDrawingCreatorState extends DrawingCreatorState<Drawing> {
   /// If drawing has been started.
   bool _isPenDown = false;
 
@@ -56,7 +68,7 @@ class _LineDrawingCreatorState extends DrawingCreatorState<LineDrawing> {
         ));
         _isPenDown = true;
 
-        drawingParts.add(LineDrawing(
+        drawingParts.add((widget as LineDrawingCreator).createLineDrawing(
           drawingPart: DrawingParts.marker,
           startEdgePoint: edgePoints.first,
         ));
@@ -82,12 +94,12 @@ class _LineDrawingCreatorState extends DrawingCreatorState<LineDrawing> {
         } else {
           /// If the initial point and the final point are not the same,
           /// draw the final point and the whole line.
-          drawingParts.addAll(<LineDrawing>[
-            LineDrawing(
+          drawingParts.addAll(<Drawing>[
+            (widget as LineDrawingCreator).createLineDrawing(
               drawingPart: DrawingParts.marker,
               endEdgePoint: edgePoints[currentTap],
             ),
-            LineDrawing(
+            (widget as LineDrawingCreator).createLineDrawing(
               drawingPart: DrawingParts.line,
               startEdgePoint: edgePoints[previousTap],
               endEdgePoint: edgePoints[currentTap],
