@@ -31,6 +31,49 @@ void paintDot(
         ..strokeWidth = 1,
     );
 
+/// Paints a dot at [center]. If [hasGlow] is true, a semi-transparent glow is
+/// drawn around the dot. [dotRadius] and [glowRadius] control their sizes.
+void paintDotWithGlow(
+  Canvas canvas,
+  Offset center, {
+  Paint? paint,
+  double dotRadius = 3,
+  Color? color,
+  bool hasGlow = false,
+  double? glowRadius,
+  bool visible = true,
+}) {
+  if (!visible) {
+    return;
+  }
+
+  // Use default paint if none provided
+  paint ??= Paint()
+    ..style = PaintingStyle.fill
+    ..strokeWidth = 1;
+
+  // Set color if specified
+  if (color != null) {
+    paint.color = color;
+  }
+
+  // Draw the dot
+  canvas.drawCircle(center, dotRadius, paint);
+
+  if (hasGlow) {
+    // Set glow radius, default to 2x dot size if not specified
+    glowRadius ??= dotRadius * 2;
+
+    // // Create a semi-transparent glow paint
+    final Paint glowPaint = Paint()
+      ..color = paint.color.withOpacity(0.5)
+      ..style = PaintingStyle.fill;
+
+    // Draw the glow
+    canvas.drawCircle(center, glowRadius, glowPaint);
+  }
+}
+
 /// Paints a blinking dot on [center].
 void paintBlinkingDot(
   Canvas canvas,
@@ -39,7 +82,7 @@ void paintBlinkingDot(
   AnimationInfo animationInfo,
   Color color,
 ) {
-  paintDot(canvas, Offset(dotX, y), color);
+  paintDotWithGlow(canvas, Offset(dotX, y), color: color);
 
   paintBlinkingGlow(
     canvas,
